replace notebook with main component
